requirements.txt
aiogram==3.4.1
Flask==3.0.0
python-dotenv==1.0.1